AI-Foundry-Agent-App-Using-MCP/requirements.txt
python-dotenv~=1.0.0
azure-identity>=1.15.0
azure-ai-projects
azure-ai-agents>=1.1.0b4